requirements.txt
Algorithmia
requests
pyyaml
playsound